feat : CommentBox 컴포넌트화
