--- frontend/components/Comment.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const Comment = () => {
-  return <Box></Box>;
-};
-export default Comment;
-
-const Box = styled.li``;
