+++ frontend/components/CommentBox.tsx
@@ -1,0 +1,117 @@
+import styled from 'styled-components';
+import EiditorSkeleton from './skeleton/EiditorSkeleton';
+import dynamic from 'next/dynamic';
+import hljs from 'highlight.js';
+import Pagenation from './Pagenation';
+
+const Editor = dynamic(() => import('@/components/editor/Editor'), {
+  ssr: false,
+  loading: () => <EiditorSkeleton />,
+});
+
+//이상하게 Editor에서 조건부로 옵션을 설정하면 editor가 고장나서 상위에서 설정한 옵션을 내려주는 방식으로 해결하였음
+const COMMENT_OPTIONS: EasyMDE.Options = {
+  renderingConfig: {
+    codeSyntaxHighlighting: true,
+    hljs,
+  }, //hljs 사용
+  maxHeight: '120px',
+  spellChecker: false, //스펠체크 off
+  status: false, //우측 하단 상태
+  previewClass: ['markdown-body'], //github 마크다운 사용
+  hideIcons: ['guide', 'fullscreen', 'side-by-side'], //버튼 가리기
+};
+
+type Props = {
+  commentData: string[];
+  addComment: () => void;
+  commentVal: string;
+  changeCommentVal: (value: string) => void;
+  commentPageHandler: (page: number) => void;
+  commentPage: number;
+};
+
+const CommentBox = ({
+  commentData,
+  addComment,
+  commentVal,
+  changeCommentVal,
+  commentPageHandler,
+  commentPage,
+}: Props) => {
+  return (
+    <Box>
+      <div className="comment-write-box">
+        <div className="comment-submit-box">
+          <button onClick={addComment}>댓글 작성</button>
+        </div>
+        <Editor
+          content={commentVal}
+          commentOptions={COMMENT_OPTIONS}
+          changeContent={changeCommentVal}
+          type={'comment'}
+        />
+      </div>
+      <div className="view-comment">
+        <ul>
+          {commentData.map((comment, i) => (
+            <li className="comment" key={`${comment}+${i}`}>
+              {comment}
+            </li>
+          ))}
+        </ul>
+      </div>
+      {commentData.length > 0 && (
+        <Pagenation
+          onPageChange={commentPageHandler}
+          page={commentPage}
+          pageSize={Math.ceil(commentData.length / 5)} //서버 데이터로 변경해야함
+        />
+      )}
+    </Box>
+  );
+};
+export default CommentBox;
+
+const Box = styled.div`
+  display: flex;
+  flex-direction: column;
+  font-size: 15px;
+  > div {
+    width: 100%;
+  }
+
+  .view-comment {
+    ul {
+      display: flex;
+      flex-direction: column;
+      gap: 16px;
+      li {
+      }
+    }
+  }
+
+  .comment-write-box {
+    width: 100%;
+    position: relative;
+    .comment-submit-box {
+      position: absolute;
+      top: 6px;
+      z-index: 2;
+      right: 6px;
+      button {
+        height: 100%;
+        padding: 10px 10px;
+        font-size: 14px;
+        border: none;
+        cursor: pointer;
+        transition: background 0.5s ease, color 0.5s ease;
+        :hover {
+          background: #9b7aff;
+          color: white;
+          border-radius: 5px;
+        }
+      }
+    }
+  }
+`;
