+++ frontend/pages/project/[id]/index.tsx
@@ -21,6 +21,7 @@ import hljs from 'highlight.js';
 import EiditorSkeleton from '@/components/skeleton/EiditorSkeleton';
 import Pagenation from '@/components/Pagenation';
 import { BUTTON_STATE } from '@/constant/constant';
+import CommentBox from '@/components/CommentBox';
 const ReactMarkdown = dynamic(() => import('@/components/editor/ContentBox'), {
   ssr: false,
   loading: () => <ContentSkeleton />,
@@ -268,35 +269,14 @@ const ViewProject = () => {
               <span>{data.totalLikes}</span>
             </div>
           </div>
-          <div className="comment-box">
-            <div className="view-comment">
-              <ul>
-                {commentData.map((comment, i) => (
-                  <li className="comment" key={`${comment}+${i}`}>
-                    {comment}
-                  </li>
-                ))}
-              </ul>
-            </div>
-            {commentData.length > 0 && (
-              <Pagenation
-                onPageChange={commentPageHandler}
-                page={commentPage}
-                pageSize={Math.ceil(commentData.length / 5)} //서버 데이터로 변경해야함
-              />
-            )}
-            <div className="comment-write-box">
-              <div className="comment-submit-box">
-                <button onClick={addComment}>댓글 작성</button>
-              </div>
-              <Editor
-                content={commentVal}
-                commentOptions={COMMENT_OPTIONS}
-                changeContent={changeCommentVal}
-                type={'comment'}
-              />
-            </div>
-          </div>
+          <CommentBox
+            commentData={commentData}
+            addComment={addComment}
+            commentVal={commentVal}
+            changeCommentVal={changeCommentVal}
+            commentPageHandler={commentPageHandler}
+            commentPage={commentPage}
+          />
         </Main>
       </GridBox>
     );
@@ -309,49 +289,6 @@ const Main = styled.div`
   display: flex;
   flex-direction: column;
   gap: 32px;
-
-  .comment-box {
-    display: flex;
-    flex-direction: column;
-    font-size: 15px;
-    > div {
-      width: 100%;
-    }
-
-    .view-comment {
-      ul {
-        display: flex;
-        flex-direction: column;
-        gap: 16px;
-        li {
-        }
-      }
-    }
-
-    .comment-write-box {
-      width: 100%;
-      position: relative;
-      .comment-submit-box {
-        position: absolute;
-        top: 6px;
-        z-index: 2;
-        right: 6px;
-        button {
-          height: 100%;
-          padding: 10px 10px;
-          font-size: 14px;
-          border: none;
-          cursor: pointer;
-          transition: background 0.5s ease, color 0.5s ease;
-          :hover {
-            background: #9b7aff;
-            color: white;
-            border-radius: 5px;
-          }
-        }
-      }
-    }
-  }
 
   .title {
     display: flex;
